Corrigindo os code smells Long Method do signUp_controller e duplicate code do ods.dart
