--- lib/services/ods.dart
@@ -1,20 +1,0 @@
-class ODS {
-  static const List<String> all_ods = [
-    "Erradicação da Pobreza",
-    "Fome Zero e Agricultura Sustentável",
-    "Saúde e Bem-Estar",
-    "Educação de Qualidade",
-    "Igualdade de Gênero",
-    "Água Potável e Saneamento",
-    "Energia Acessível e Limpa", 
-    "Trabalho Decente e Crescimento Econômico",
-    "Indústria, Inovação e Infraestrutura",
-    "Redução das Desigualdades",
-    "Cidades e Comunidades Sustentáveis",
-    "Consumo e Produção Responsáveis",
-    "Ação Contra a Mudança Global do Clima",
-    "Vida na Água",
-    "Vida Terrestre",
-    "Paz, Justiça e Instituições Eficazes",
-  ];
-}
